+++ plugins/modules/oci_database_autonomous_container_database.py
@@ -36,8 +36,8 @@ options:
         type: str
     db_name:
         description:
-            - The database name for the Autonomous Container Database. The name must be unique within the Cloud Autonomous VM Cluster, must start with an
-              alphabetic character and followed by 1 to 7 alphanumeric characters.
+            - The Database name for the Autonomous Container Database. The name must be unique within the Cloud Autonomous VM Cluster, starting with an
+              alphabetic character, followed by 1 to 7 alphanumeric characters.
         type: str
     service_level_agreement_type:
         description:
@@ -606,8 +606,8 @@ autonomous_container_database:
             sample: db_unique_name_example
         db_name:
             description:
-                - The database name for the Autonomous Container Database. The name must be unique within the Cloud Autonomous VM Cluster, must start with an
-                  alphabetic character and followed by 1 to 7 alphanumeric characters.
+                - The Database name for the Autonomous Container Database. The name must be unique within the Cloud Autonomous VM Cluster, starting with an
+                  alphabetic character, followed by 1 to 7 alphanumeric characters.
             returned: on success
             type: str
             sample: db_name_example