+++ plugins/modules/oci_database_autonomous_vm_cluster_facts.py
@@ -201,7 +201,8 @@ autonomous_vm_clusters:
             sample: 56
         memory_per_oracle_compute_unit_in_gbs:
             description:
-                - The amount of memory (in GBs) to be enabled per each CPU core.
+                - The amount of memory (in GBs) to be enabled per OCPU or ECPU. See L(Compute Models in Autonomous Database on Dedicated Exadata
+                  Infrastructure,https://docs.oracle.com/en/cloud/paas/autonomous-database/dedicated/adbak) for more details.
             returned: on success
             type: int
             sample: 56
@@ -425,13 +426,13 @@ autonomous_vm_clusters:
             sample: true
         time_database_ssl_certificate_expires:
             description:
-                - The date and time of Database SSL certificate expiration.
+                - The date and time of the Database SSL certificate expiration.
             returned: on success
             type: str
             sample: "2013-10-20T19:20:30+01:00"
         time_ords_certificate_expires:
             description:
-                - The date and time of ORDS certificate expiration.
+                - The date and time of the ORDS certificate expiration.
             returned: on success
             type: str
             sample: "2013-10-20T19:20:30+01:00"