+++ plugins/modules/oci_database_db_system_shape_facts.py
@@ -72,8 +72,8 @@ db_system_shapes:
             sample: shape_family_example
         shape_type:
             description:
-                - The shape type for the virtual machine DB system. Shape type is determined by CPU hardware. Valid values are `AMD` , `INTEL` or
-                  `INTEL_FLEX_X9`.
+                - The shape type for the virtual machine DB system. Shape type is determined by CPU hardware. Valid values are `AMD` , `INTEL`, `INTEL_FLEX_X9`
+                  or `AMPERE_FLEX_A1`.
             returned: on success
             type: str
             sample: AMD